feat: Moved mol parser tests to tests folder
