+++ tests/test_parsers/test_mol_parser.py
@@ -1,0 +1,3437 @@
+import os
+import pytest
+
+import numpy as np
+
+from manim_chemistry.utils import MolParser
+
+base_files_path = os.path.join("examples", "element_files")
+
+
+class TestMolParser:
+    morphine_mol_file = os.path.join(base_files_path, "morphine.mol")
+    morphine_3d_mol_file = os.path.join(base_files_path, "morphine3d.mol")
+    heme_2d_mol_file = os.path.join(base_files_path, "heme_2d.mol")
+    heme_group_mol_file = os.path.join(base_files_path, "heme_group.mol")
+
+    @pytest.fixture
+    def morphine_file_data(self):
+        return [
+            "\n",
+            "  Marvin  02020822302D          \n",
+            "\n",
+            " 22 26  0  0  1  0            999 V2000\n",
+            "    0.0000   -0.8250    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   -0.7145   -0.4125    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "    0.7145   -0.4125    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   -0.7145    0.4125    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "    0.7145    0.4125    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "    0.0000    0.8250    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "    0.0030   -1.6464    0.0000 C   0  0  2  0  0  0  0  0  0  0  0  0\n",
+            "    0.7109   -2.0631    0.0000 C   0  0  2  0  0  0  0  0  0  0  0  0\n",
+            "   -0.7127   -2.0511    0.0000 C   0  0  1  0  0  0  0  0  0  0  0  0\n",
+            "    1.4268   -1.6510    0.0000 C   0  0  1  0  0  0  0  0  0  0  0  0\n",
+            "    0.7067   -2.8808    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   -1.5371   -1.1745    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   -0.7206   -2.8689    0.0000 C   0  0  2  0  0  0  0  0  0  0  0  0\n",
+            "    1.4310   -0.8297    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "    2.1419   -2.0715    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   -0.0128   -3.2891    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   -1.4365   -3.2845    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   -1.4262    0.8296    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "    2.9416   -1.8530    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "    1.4954   -2.6258    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "    2.1397   -1.2341    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "    0.7183   -1.2330    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "  7  1  1  0  0  0  0\n",
+            "  1  2  1  0  0  0  0\n",
+            "  2 12  1  0  0  0  0\n",
+            "  1  3  2  0  0  0  0\n",
+            "  3 14  1  0  0  0  0\n",
+            "  2  4  2  0  0  0  0\n",
+            "  4 18  1  0  0  0  0\n",
+            "  3  5  1  0  0  0  0\n",
+            "  4  6  1  0  0  0  0\n",
+            "  5  6  2  0  0  0  0\n",
+            "  7  8  1  0  0  0  0\n",
+            "  7  9  1  0  0  0  0\n",
+            "  7 22  1  1  0  0  0\n",
+            "  8 10  1  0  0  0  0\n",
+            "  8 11  1  0  0  0  0\n",
+            "  8 20  1  1  0  0  0\n",
+            "  9 12  1  6  0  0  0\n",
+            "  9 13  1  0  0  0  0\n",
+            " 10 15  1  1  0  0  0\n",
+            " 10 14  1  0  0  0  0\n",
+            " 11 16  2  0  0  0  0\n",
+            " 13 17  1  6  0  0  0\n",
+            " 13 16  1  0  0  0  0\n",
+            " 15 19  1  0  0  0  0\n",
+            " 21 15  1  0  0  0  0\n",
+            " 22 21  1  0  0  0  0\n",
+            "M  END\n",
+        ]
+
+    @pytest.fixture
+    def morphine_3d_file_data(self):
+        return [
+            "\n",
+            "  ACD/LABS04182319523D\n",
+            "\n",
+            " 40 44  0  0  1  0  0  0  0  0  1 V2000\n",
+            "   18.7847   -6.4578   -2.0216 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.4847   -6.0807   -1.8514 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.6943   -5.7779   -2.7745 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.0681   -4.8705   -2.3788 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.2967   -4.5612   -3.3125 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.9907   -4.1082   -3.0991 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.0067   -7.7069   -1.3775 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.7749   -8.6094   -2.3038 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.6136   -8.2021   -1.0412 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.1805   -7.9667   -2.5763 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.9955   -8.7657   -3.5788 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.7395   -7.0476   -1.1824 O   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.1717   -9.3637   -1.9997 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.0493   -6.4355   -3.0215 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.9630   -8.0363   -1.3315 N   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.7119   -9.1268   -3.4162 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   15.7548   -9.4200   -2.0456 O   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   15.7881   -4.4406   -2.2167 O   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   23.3571   -7.6257   -1.5536 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.9185   -9.6121   -1.8396 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.3225   -7.1799   -0.3121 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.8476   -7.5860   -0.0787 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.5788   -8.5994    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.6885   -8.5338   -3.3906 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.5755  -10.3275   -1.6103 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.0152   -3.9572   -3.9068 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.6800   -3.1237   -3.5096 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.3600   -8.6178   -4.6179 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.2458   -6.3601   -4.1124 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.8010   -5.8341   -2.4668 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.1473   -9.2211   -4.3683 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   15.2717  -10.2576   -1.7471 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   15.4952   -3.5598   -2.6199 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   23.5561   -7.5671   -2.6451 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   23.5273   -6.6294   -1.0922 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   24.0402   -8.3708   -1.0922 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.8979   -7.2823    0.6327 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.3745   -6.1280   -0.6659 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.3757   -6.8198    0.5728 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.8362   -8.5697    0.4375 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "  7  1  1  0  0  0  0\n",
+            "  1  2  1  0  0  0  0\n",
+            "  2 12  1  0  0  0  0\n",
+            "  1  3  2  0  0  0  0\n",
+            "  3 14  1  0  0  0  0\n",
+            "  2  4  2  0  0  0  0\n",
+            "  4 18  1  0  0  0  0\n",
+            "  3  5  1  0  0  0  0\n",
+            "  4  6  1  0  0  0  0\n",
+            "  5  6  2  0  0  0  0\n",
+            "  7  8  1  0  0  0  0\n",
+            "  7  9  1  0  0  0  0\n",
+            "  7 22  1  1  0  0  0\n",
+            "  8 10  1  0  0  0  0\n",
+            "  8 11  1  0  0  0  0\n",
+            "  8 20  1  1  0  0  0\n",
+            "  9 12  1  6  0  0  0\n",
+            "  9 13  1  0  0  0  0\n",
+            " 10 15  1  1  0  0  0\n",
+            " 10 14  1  0  0  0  0\n",
+            " 11 16  2  0  0  0  0\n",
+            " 13 17  1  6  0  0  0\n",
+            " 13 16  1  0  0  0  0\n",
+            " 15 19  1  0  0  0  0\n",
+            " 21 15  1  0  0  0  0\n",
+            " 22 21  1  0  0  0  0\n",
+            " 23  9  1  0  0  0  0\n",
+            " 24 10  1  0  0  0  0\n",
+            " 25 13  1  0  0  0  0\n",
+            " 26  5  1  0  0  0  0\n",
+            " 27  6  1  0  0  0  0\n",
+            " 28 11  1  0  0  0  0\n",
+            " 29 14  1  0  0  0  0\n",
+            " 30 14  1  0  0  0  0\n",
+            " 31 16  1  0  0  0  0\n",
+            " 32 17  1  0  0  0  0\n",
+            " 33 18  1  0  0  0  0\n",
+            " 34 19  1  0  0  0  0\n",
+            " 35 19  1  0  0  0  0\n",
+            " 36 19  1  0  0  0  0\n",
+            " 37 21  1  0  0  0  0\n",
+            " 38 21  1  0  0  0  0\n",
+            " 39 22  1  0  0  0  0\n",
+            " 40 22  1  0  0  0  0\n",
+            "M  END\n",
+        ]
+
+    @pytest.fixture
+    def heme_2d_file_data(self):
+        return [
+            "\n",
+            "  ACD/LABS04222323523D\n",
+            "\n",
+            " 73 76  0  0  0  0  0  0  0  0  2 V2000\n",
+            "   19.8676  -19.1177   -1.5852 O   0  5  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.6176  -17.7372   -1.5534 O   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   27.7926  -11.3449   -1.3889 O   0  5  0  0  0  0  0  0  0  0  0  0\n",
+            "   26.2694  -12.9697   -1.4323 O   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.7120  -11.9403   -1.2028 N   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.6948   -9.9176   -1.2420 N   0  5  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.6792   -9.9261   -1.2508 N   0  5  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.6520   -7.9140   -1.2977 N   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.7942  -12.7420   -1.1871 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.5042  -10.9948   -1.2120 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.6235  -12.7312   -1.1933 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.4739   -8.8181   -1.2818 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   15.8963  -11.0195   -1.2741 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.7297   -7.1105   -1.3373 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.5621   -7.1222   -1.3162 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   15.8641   -8.8548   -1.3049 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.1173  -12.3295   -1.1815 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.3583  -14.1568   -1.1795 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   22.8218  -10.5756   -1.2277 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.3017  -12.3365   -1.2347 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.0557   -7.4979   -1.3319 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.0198  -14.1321   -1.1515 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   22.8005   -9.1973   -1.2723 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.2416   -7.5280   -1.3195 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   14.5740  -10.6443   -1.3497 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.9666   -5.7197   -1.3620 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.3008   -5.7093   -1.3935 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   14.5546   -9.2737   -1.3631 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.2863  -15.3740   -1.1968 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   24.0278  -11.5107   -1.2136 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.0435  -15.2964   -1.0919 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   23.9858   -8.2439   -1.3163 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   13.3797  -11.5726   -1.4182 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.0243   -4.5359   -1.4015 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.2129   -4.5454   -1.5006 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   13.3706   -8.4010   -1.4492 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.5448  -16.7096   -1.1026 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   25.3819  -10.8068   -1.1561 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.0311   -3.4198   -0.7937 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   12.2878   -8.6418   -0.6967 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.4233  -17.8843   -1.4349 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   26.5085  -11.7886   -1.3281 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.9066  -13.1092   -1.1619 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   15.5193  -13.1211   -1.2439 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.8243   -6.6991   -1.3699 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   15.4484   -6.7533   -1.3538 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.8530  -15.3637   -2.1563 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.9975  -15.2995   -0.3446 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   23.9479  -12.1771   -0.3268 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   23.9942  -12.1286   -2.1377 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.0463  -14.9562   -0.7387 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.9377  -15.7260   -2.1108 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.4003  -16.0738   -0.3855 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   24.7924   -8.5743   -0.6292 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   24.3704   -8.2067   -2.3585 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   23.6777   -7.2237   -1.0041 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   12.5446  -11.0758   -1.9545 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   13.0626  -11.8208   -0.3832 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   13.6319  -12.5121   -1.9593 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.4405   -3.6831   -0.8297 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.8732   -4.2338   -2.4592 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.0435   -4.8051   -0.9507 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.0809   -4.6133   -2.1892 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   13.3792   -7.5400   -2.1494 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.7066  -16.6943   -1.8320 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.1517  -16.8412   -0.0710 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   25.4896  -10.3254   -0.1583 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   25.4419  -10.0395   -1.9593 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.1774   -3.3276   -0.0927 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.7461   -2.5757   -0.9082 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   12.2820   -9.5056    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   11.3978   -7.9789   -0.7703 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.5078  -10.0631    0.0000 Fe  0  2  0  0  0  0  0  0  0  0  0  0\n",
+            "  1 41  1  0  0  0  0\n",
+            "  2 41  2  0  0  0  0\n",
+            "  3 42  1  0  0  0  0\n",
+            "  4 42  2  0  0  0  0\n",
+            "  5  9  2  0  0  0  0\n",
+            "  5 11  1  0  0  0  0\n",
+            "  6 10  1  0  0  0  0\n",
+            "  6 12  1  0  0  0  0\n",
+            "  7 13  1  0  0  0  0\n",
+            "  7 16  1  0  0  0  0\n",
+            "  8 14  2  0  0  0  0\n",
+            "  8 15  1  0  0  0  0\n",
+            "  9 17  1  0  0  0  0\n",
+            "  9 18  1  0  0  0  0\n",
+            " 10 17  2  0  0  0  0\n",
+            " 10 19  1  0  0  0  0\n",
+            " 11 20  2  0  0  0  0\n",
+            " 11 22  1  0  0  0  0\n",
+            " 12 21  2  0  0  0  0\n",
+            " 12 23  1  0  0  0  0\n",
+            " 13 20  1  0  0  0  0\n",
+            " 13 25  2  0  0  0  0\n",
+            " 14 21  1  0  0  0  0\n",
+            " 14 27  1  0  0  0  0\n",
+            " 15 24  2  0  0  0  0\n",
+            " 15 26  1  0  0  0  0\n",
+            " 16 24  1  0  0  0  0\n",
+            " 16 28  2  0  0  0  0\n",
+            " 17 43  1  0  0  0  0\n",
+            " 18 22  2  0  0  0  0\n",
+            " 18 29  1  0  0  0  0\n",
+            " 19 23  2  0  0  0  0\n",
+            " 19 30  1  0  0  0  0\n",
+            " 20 44  1  0  0  0  0\n",
+            " 21 45  1  0  0  0  0\n",
+            " 22 31  1  0  0  0  0\n",
+            " 23 32  1  0  0  0  0\n",
+            " 24 46  1  0  0  0  0\n",
+            " 25 28  1  0  0  0  0\n",
+            " 25 33  1  0  0  0  0\n",
+            " 26 27  2  0  0  0  0\n",
+            " 26 34  1  0  0  0  0\n",
+            " 27 35  1  0  0  0  0\n",
+            " 28 36  1  0  0  0  0\n",
+            " 29 37  1  0  0  0  0\n",
+            " 29 47  1  0  0  0  0\n",
+            " 29 48  1  0  0  0  0\n",
+            " 30 38  1  0  0  0  0\n",
+            " 30 49  1  0  0  0  0\n",
+            " 30 50  1  0  0  0  0\n",
+            " 31 51  1  0  0  0  0\n",
+            " 31 52  1  0  0  0  0\n",
+            " 31 53  1  0  0  0  0\n",
+            " 32 54  1  0  0  0  0\n",
+            " 32 55  1  0  0  0  0\n",
+            " 32 56  1  0  0  0  0\n",
+            " 33 57  1  0  0  0  0\n",
+            " 33 58  1  0  0  0  0\n",
+            " 33 59  1  0  0  0  0\n",
+            " 34 60  1  0  0  0  0\n",
+            " 34 61  1  0  0  0  0\n",
+            " 34 62  1  0  0  0  0\n",
+            " 35 39  2  0  0  0  0\n",
+            " 35 63  1  0  0  0  0\n",
+            " 36 40  2  0  0  0  0\n",
+            " 36 64  1  0  0  0  0\n",
+            " 37 41  1  0  0  0  0\n",
+            " 37 65  1  0  0  0  0\n",
+            " 37 66  1  0  0  0  0\n",
+            " 38 42  1  0  0  0  0\n",
+            " 38 67  1  0  0  0  0\n",
+            " 38 68  1  0  0  0  0\n",
+            " 39 69  1  0  0  0  0\n",
+            " 39 70  1  0  0  0  0\n",
+            " 40 71  1  0  0  0  0\n",
+            " 40 72  1  0  0  0  0\n",
+            "M  CHG  5   1  -1   3  -1   6  -1   7  -1  73   2\n",
+            "M  END\n",
+        ]
+
+    @pytest.fixture
+    def heme_group_file_data(self):
+        return [
+            "\n",
+            "  ACD/LABS04222323342D\n",
+            "\n",
+            " 73 76  0  0  0  0  0  0  0  0  2 V2000\n",
+            "   18.8560  -10.1645    0.0000 Fe  0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.4111  -18.3615    0.0000 O   0  5  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.4059  -17.2096    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   27.1847  -11.2000    0.0000 O   0  5  0  0  0  0  0  0  0  0  0  0\n",
+            "   25.7979  -13.0395    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.7949  -11.5492    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.5798  -10.0015    0.0000 N   0  5  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.9864  -10.0506    0.0000 N   0  5  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.7584   -8.4699    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.7909  -12.5337    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.4317  -10.9754    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.7681  -12.5760    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.3894   -9.0370    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.1676  -11.0177    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.7909   -7.4787    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.7258   -7.5210    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.1676   -9.0793    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.0536  -12.1959    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.4109  -13.7543    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   22.6120  -10.6375    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.5476  -12.1959    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.0115   -7.8588    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.1482  -13.7543    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   22.6120   -9.3748    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.5054   -7.8588    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   14.9472  -10.6797    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.1059   -6.3004    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.3686   -6.3004    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   14.9472   -9.4171    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.0759  -14.9060    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   23.6739  -11.4382    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.3654  -14.8294    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   23.6703   -8.5694    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   13.8886  -11.4850    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.3176   -5.2292    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.1343   -5.2129    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   13.8886   -8.6118    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.4109  -16.0578    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   24.8983  -10.9187    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   19.5752   -4.0060    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   12.6620   -9.1259    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.0759  -17.2096    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   25.9603  -11.7195    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.6253  -12.7901    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   15.9674  -12.7819    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   21.5920   -7.2733    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   15.9223   -7.2757    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.7077  -15.4361    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.7077  -14.3760    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   23.0707  -12.0004    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   24.1229  -12.1298    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.6988  -14.3440    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.8799  -15.4960    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.0320  -15.3148    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   24.1697   -9.2256    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   24.3265   -8.0700    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   23.1708   -7.9133    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   13.3894  -10.8288    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   13.2323  -11.9843    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   14.3879  -12.1412    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   17.9819   -4.7405    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.8289   -4.5650    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   16.6536   -5.7180    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.9556   -5.2868    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   13.9929   -7.7939    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.7793  -15.5278    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.7793  -16.5878    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   25.5015  -10.3565    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   24.4494  -10.2271    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   20.0499   -3.3318    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   18.7539   -3.9322    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   12.0058   -8.6267    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "   12.5577   -9.9439    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n",
+            "  2 42  1  0  0  0  0\n",
+            "  3 42  2  0  0  0  0\n",
+            "  4 43  1  0  0  0  0\n",
+            "  5 43  2  0  0  0  0\n",
+            "  6 10  2  0  0  0  0\n",
+            "  6 12  1  0  0  0  0\n",
+            "  7 11  1  0  0  0  0\n",
+            "  7 13  1  0  0  0  0\n",
+            "  8 14  1  0  0  0  0\n",
+            "  8 17  1  0  0  0  0\n",
+            "  9 15  2  0  0  0  0\n",
+            "  9 16  1  0  0  0  0\n",
+            " 10 18  1  0  0  0  0\n",
+            " 10 19  1  0  0  0  0\n",
+            " 11 18  2  0  0  0  0\n",
+            " 11 20  1  0  0  0  0\n",
+            " 12 21  2  0  0  0  0\n",
+            " 12 23  1  0  0  0  0\n",
+            " 13 22  2  0  0  0  0\n",
+            " 13 24  1  0  0  0  0\n",
+            " 14 21  1  0  0  0  0\n",
+            " 14 26  2  0  0  0  0\n",
+            " 15 22  1  0  0  0  0\n",
+            " 15 28  1  0  0  0  0\n",
+            " 16 25  2  0  0  0  0\n",
+            " 16 27  1  0  0  0  0\n",
+            " 17 25  1  0  0  0  0\n",
+            " 17 29  2  0  0  0  0\n",
+            " 18 44  1  0  0  0  0\n",
+            " 19 23  2  0  0  0  0\n",
+            " 19 30  1  0  0  0  0\n",
+            " 20 24  2  0  0  0  0\n",
+            " 20 31  1  0  0  0  0\n",
+            " 21 45  1  0  0  0  0\n",
+            " 22 46  1  0  0  0  0\n",
+            " 23 32  1  0  0  0  0\n",
+            " 24 33  1  0  0  0  0\n",
+            " 25 47  1  0  0  0  0\n",
+            " 26 29  1  0  0  0  0\n",
+            " 26 34  1  0  0  0  0\n",
+            " 27 28  2  0  0  0  0\n",
+            " 27 35  1  0  0  0  0\n",
+            " 28 36  1  0  0  0  0\n",
+            " 29 37  1  0  0  0  0\n",
+            " 30 38  1  0  0  0  0\n",
+            " 30 48  1  0  0  0  0\n",
+            " 30 49  1  0  0  0  0\n",
+            " 31 39  1  0  0  0  0\n",
+            " 31 50  1  0  0  0  0\n",
+            " 31 51  1  0  0  0  0\n",
+            " 32 52  1  0  0  0  0\n",
+            " 32 53  1  0  0  0  0\n",
+            " 32 54  1  0  0  0  0\n",
+            " 33 55  1  0  0  0  0\n",
+            " 33 56  1  0  0  0  0\n",
+            " 33 57  1  0  0  0  0\n",
+            " 34 58  1  0  0  0  0\n",
+            " 34 59  1  0  0  0  0\n",
+            " 34 60  1  0  0  0  0\n",
+            " 35 61  1  0  0  0  0\n",
+            " 35 62  1  0  0  0  0\n",
+            " 35 63  1  0  0  0  0\n",
+            " 36 40  2  0  0  0  0\n",
+            " 36 64  1  0  0  0  0\n",
+            " 37 41  2  0  0  0  0\n",
+            " 37 65  1  0  0  0  0\n",
+            " 38 42  1  0  0  0  0\n",
+            " 38 66  1  0  0  0  0\n",
+            " 38 67  1  0  0  0  0\n",
+            " 39 43  1  0  0  0  0\n",
+            " 39 68  1  0  0  0  0\n",
+            " 39 69  1  0  0  0  0\n",
+            " 40 70  1  0  0  0  0\n",
+            " 40 71  1  0  0  0  0\n",
+            " 41 72  1  0  0  0  0\n",
+            " 41 73  1  0  0  0  0\n",
+            "M  CHG  4   2  -1   4  -1   7  -1   8  -1\n",
+            "M  END\n",
+        ]
+
+    @pytest.fixture
+    def morphine_atoms_data(self):
+        return {
+            1: {
+                "coords": np.array([0.0, -0.825, 0.0]),
+                "element": "C",
+                "bond_to": {7: "C", 2: "C", 3: "C"},
+            },
+            2: {
+                "coords": np.array([-0.7145, -0.4125, 0.0]),
+                "element": "C",
+                "bond_to": {1: "C", 12: "O", 4: "C"},
+            },
+            3: {
+                "coords": np.array([0.7145, -0.4125, 0.0]),
+                "element": "C",
+                "bond_to": {1: "C", 14: "C", 5: "C"},
+            },
+            4: {
+                "coords": np.array([-0.7145, 0.4125, 0.0]),
+                "element": "C",
+                "bond_to": {2: "C", 18: "O", 6: "C"},
+            },
+            5: {
+                "coords": np.array([0.7145, 0.4125, 0.0]),
+                "element": "C",
+                "bond_to": {3: "C", 6: "C"},
+            },
+            6: {
+                "coords": np.array([0.0, 0.825, 0.0]),
+                "element": "C",
+                "bond_to": {4: "C", 5: "C"},
+            },
+            7: {
+                "coords": np.array([0.003, -1.6464, 0.0]),
+                "element": "C",
+                "bond_to": {1: "C", 8: "C", 9: "C", 22: "C"},
+            },
+            8: {
+                "coords": np.array([0.7109, -2.0631, 0.0]),
+                "element": "C",
+                "bond_to": {7: "C", 10: "C", 11: "C", 20: "H"},
+            },
+            9: {
+                "coords": np.array([-0.7127, -2.0511, 0.0]),
+                "element": "C",
+                "bond_to": {7: "C", 12: "O", 13: "C"},
+            },
+            10: {
+                "coords": np.array([1.4268, -1.651, 0.0]),
+                "element": "C",
+                "bond_to": {8: "C", 15: "N", 14: "C"},
+            },
+            11: {
+                "coords": np.array([0.7067, -2.8808, 0.0]),
+                "element": "C",
+                "bond_to": {8: "C", 16: "C"},
+            },
+            12: {
+                "coords": np.array([-1.5371, -1.1745, 0.0]),
+                "element": "O",
+                "bond_to": {2: "C", 9: "C"},
+            },
+            13: {
+                "coords": np.array([-0.7206, -2.8689, 0.0]),
+                "element": "C",
+                "bond_to": {9: "C", 17: "O", 16: "C"},
+            },
+            14: {
+                "coords": np.array([1.431, -0.8297, 0.0]),
+                "element": "C",
+                "bond_to": {3: "C", 10: "C"},
+            },
+            15: {
+                "coords": np.array([2.1419, -2.0715, 0.0]),
+                "element": "N",
+                "bond_to": {10: "C", 19: "C", 21: "C"},
+            },
+            16: {
+                "coords": np.array([-0.0128, -3.2891, 0.0]),
+                "element": "C",
+                "bond_to": {11: "C", 13: "C"},
+            },
+            17: {
+                "coords": np.array([-1.4365, -3.2845, 0.0]),
+                "element": "O",
+                "bond_to": {13: "C"},
+            },
+            18: {
+                "coords": np.array([-1.4262, 0.8296, 0.0]),
+                "element": "O",
+                "bond_to": {4: "C"},
+            },
+            19: {
+                "coords": np.array([2.9416, -1.853, 0.0]),
+                "element": "C",
+                "bond_to": {15: "N"},
+            },
+            20: {
+                "coords": np.array([1.4954, -2.6258, 0.0]),
+                "element": "H",
+                "bond_to": {8: "C"},
+            },
+            21: {
+                "coords": np.array([2.1397, -1.2341, 0.0]),
+                "element": "C",
+                "bond_to": {15: "N", 22: "C"},
+            },
+            22: {
+                "coords": np.array([0.7183, -1.233, 0.0]),
+                "element": "C",
+                "bond_to": {7: "C", 21: "C"},
+            },
+        }
+
+    @pytest.fixture
+    def morphine_bonds_data(self):
+        return {
+            7: [
+                {
+                    "to": 1,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 8,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 9,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 22,
+                    "type": "1",
+                    "stereo": 1,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            1: [
+                {
+                    "to": 2,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 3,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            2: [
+                {
+                    "to": 12,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 4,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            3: [
+                {
+                    "to": 14,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 5,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            4: [
+                {
+                    "to": 18,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 6,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            5: [
+                {
+                    "to": 6,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            8: [
+                {
+                    "to": 10,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 11,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 20,
+                    "type": "1",
+                    "stereo": 1,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            9: [
+                {
+                    "to": 12,
+                    "type": "1",
+                    "stereo": 6,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 13,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            10: [
+                {
+                    "to": 15,
+                    "type": "1",
+                    "stereo": 1,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 14,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            11: [
+                {
+                    "to": 16,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            13: [
+                {
+                    "to": 17,
+                    "type": "1",
+                    "stereo": 6,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 16,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            15: [
+                {
+                    "to": 19,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            21: [
+                {
+                    "to": 15,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            22: [
+                {
+                    "to": 21,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+        }
+
+    @pytest.fixture
+    def morphine_3d_atoms_data(self):
+        return {
+            1: {
+                "coords": np.array([18.7847, -6.4578, -2.0216]),
+                "element": "C",
+                "bond_to": {7: "C", 2: "C", 3: "C"},
+            },
+            2: {
+                "coords": np.array([17.4847, -6.0807, -1.8514]),
+                "element": "C",
+                "bond_to": {1: "C", 12: "O", 4: "C"},
+            },
+            3: {
+                "coords": np.array([19.6943, -5.7779, -2.7745]),
+                "element": "C",
+                "bond_to": {1: "C", 14: "C", 5: "C"},
+            },
+            4: {
+                "coords": np.array([17.0681, -4.8705, -2.3788]),
+                "element": "C",
+                "bond_to": {2: "C", 18: "O", 6: "C"},
+            },
+            5: {
+                "coords": np.array([19.2967, -4.5612, -3.3125]),
+                "element": "C",
+                "bond_to": {3: "C", 6: "C", 26: "H"},
+            },
+            6: {
+                "coords": np.array([17.9907, -4.1082, -3.0991]),
+                "element": "C",
+                "bond_to": {4: "C", 5: "C", 27: "H"},
+            },
+            7: {
+                "coords": np.array([19.0067, -7.7069, -1.3775]),
+                "element": "C",
+                "bond_to": {1: "C", 8: "C", 9: "C", 22: "C"},
+            },
+            8: {
+                "coords": np.array([19.7749, -8.6094, -2.3038]),
+                "element": "C",
+                "bond_to": {7: "C", 10: "C", 11: "C", 20: "H"},
+            },
+            9: {
+                "coords": np.array([17.6136, -8.2021, -1.0412]),
+                "element": "C",
+                "bond_to": {7: "C", 12: "O", 13: "C", 23: "H"},
+            },
+            10: {
+                "coords": np.array([21.1805, -7.9667, -2.5763]),
+                "element": "C",
+                "bond_to": {8: "C", 15: "N", 14: "C", 24: "H"},
+            },
+            11: {
+                "coords": np.array([18.9955, -8.7657, -3.5788]),
+                "element": "C",
+                "bond_to": {8: "C", 16: "C", 28: "H"},
+            },
+            12: {
+                "coords": np.array([16.7395, -7.0476, -1.1824]),
+                "element": "O",
+                "bond_to": {2: "C", 9: "C"},
+            },
+            13: {
+                "coords": np.array([17.1717, -9.3637, -1.9997]),
+                "element": "C",
+                "bond_to": {9: "C", 17: "O", 16: "C", 25: "H"},
+            },
+            14: {
+                "coords": np.array([21.0493, -6.4355, -3.0215]),
+                "element": "C",
+                "bond_to": {3: "C", 10: "C", 29: "H", 30: "H"},
+            },
+            15: {
+                "coords": np.array([21.963, -8.0363, -1.3315]),
+                "element": "N",
+                "bond_to": {10: "C", 19: "C", 21: "C"},
+            },
+            16: {
+                "coords": np.array([17.7119, -9.1268, -3.4162]),
+                "element": "C",
+                "bond_to": {11: "C", 13: "C", 31: "H"},
+            },
+            17: {
+                "coords": np.array([15.7548, -9.42, -2.0456]),
+                "element": "O",
+                "bond_to": {13: "C", 32: "H"},
+            },
+            18: {
+                "coords": np.array([15.7881, -4.4406, -2.2167]),
+                "element": "O",
+                "bond_to": {4: "C", 33: "H"},
+            },
+            19: {
+                "coords": np.array([23.3571, -7.6257, -1.5536]),
+                "element": "C",
+                "bond_to": {15: "N", 34: "H", 35: "H", 36: "H"},
+            },
+            20: {
+                "coords": np.array([19.9185, -9.6121, -1.8396]),
+                "element": "H",
+                "bond_to": {8: "C"},
+            },
+            21: {
+                "coords": np.array([21.3225, -7.1799, -0.3121]),
+                "element": "C",
+                "bond_to": {15: "N", 22: "C", 37: "H", 38: "H"},
+            },
+            22: {
+                "coords": np.array([19.8476, -7.586, -0.0787]),
+                "element": "C",
+                "bond_to": {7: "C", 21: "C", 39: "H", 40: "H"},
+            },
+            23: {
+                "coords": np.array([17.5788, -8.5994, 0.0]),
+                "element": "H",
+                "bond_to": {9: "C"},
+            },
+            24: {
+                "coords": np.array([21.6885, -8.5338, -3.3906]),
+                "element": "H",
+                "bond_to": {10: "C"},
+            },
+            25: {
+                "coords": np.array([17.5755, -10.3275, -1.6103]),
+                "element": "H",
+                "bond_to": {13: "C"},
+            },
+            26: {
+                "coords": np.array([20.0152, -3.9572, -3.9068]),
+                "element": "H",
+                "bond_to": {5: "C"},
+            },
+            27: {
+                "coords": np.array([17.68, -3.1237, -3.5096]),
+                "element": "H",
+                "bond_to": {6: "C"},
+            },
+            28: {
+                "coords": np.array([19.36, -8.6178, -4.6179]),
+                "element": "H",
+                "bond_to": {11: "C"},
+            },
+            29: {
+                "coords": np.array([21.2458, -6.3601, -4.1124]),
+                "element": "H",
+                "bond_to": {14: "C"},
+            },
+            30: {
+                "coords": np.array([21.801, -5.8341, -2.4668]),
+                "element": "H",
+                "bond_to": {14: "C"},
+            },
+            31: {
+                "coords": np.array([17.1473, -9.2211, -4.3683]),
+                "element": "H",
+                "bond_to": {16: "C"},
+            },
+            32: {
+                "coords": np.array([15.2717, -10.2576, -1.7471]),
+                "element": "H",
+                "bond_to": {17: "O"},
+            },
+            33: {
+                "coords": np.array([15.4952, -3.5598, -2.6199]),
+                "element": "H",
+                "bond_to": {18: "O"},
+            },
+            34: {
+                "coords": np.array([23.5561, -7.5671, -2.6451]),
+                "element": "H",
+                "bond_to": {19: "C"},
+            },
+            35: {
+                "coords": np.array([23.5273, -6.6294, -1.0922]),
+                "element": "H",
+                "bond_to": {19: "C"},
+            },
+            36: {
+                "coords": np.array([24.0402, -8.3708, -1.0922]),
+                "element": "H",
+                "bond_to": {19: "C"},
+            },
+            37: {
+                "coords": np.array([21.8979, -7.2823, 0.6327]),
+                "element": "H",
+                "bond_to": {21: "C"},
+            },
+            38: {
+                "coords": np.array([21.3745, -6.128, -0.6659]),
+                "element": "H",
+                "bond_to": {21: "C"},
+            },
+            39: {
+                "coords": np.array([19.3757, -6.8198, 0.5728]),
+                "element": "H",
+                "bond_to": {22: "C"},
+            },
+            40: {
+                "coords": np.array([19.8362, -8.5697, 0.4375]),
+                "element": "H",
+                "bond_to": {22: "C"},
+            },
+        }
+
+    @pytest.fixture
+    def morphine_3d_bonds_data(self):
+        return {
+            7: [
+                {
+                    "to": 1,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 8,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 9,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 22,
+                    "type": "1",
+                    "stereo": 1,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            1: [
+                {
+                    "to": 2,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 3,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            2: [
+                {
+                    "to": 12,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 4,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            3: [
+                {
+                    "to": 14,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 5,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            4: [
+                {
+                    "to": 18,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 6,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            5: [
+                {
+                    "to": 6,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            8: [
+                {
+                    "to": 10,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 11,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 20,
+                    "type": "1",
+                    "stereo": 1,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            9: [
+                {
+                    "to": 12,
+                    "type": "1",
+                    "stereo": 6,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 13,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            10: [
+                {
+                    "to": 15,
+                    "type": "1",
+                    "stereo": 1,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 14,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            11: [
+                {
+                    "to": 16,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            13: [
+                {
+                    "to": 17,
+                    "type": "1",
+                    "stereo": 6,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 16,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            15: [
+                {
+                    "to": 19,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            21: [
+                {
+                    "to": 15,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            22: [
+                {
+                    "to": 21,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            23: [
+                {
+                    "to": 9,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            24: [
+                {
+                    "to": 10,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            25: [
+                {
+                    "to": 13,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            26: [
+                {
+                    "to": 5,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            27: [
+                {
+                    "to": 6,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            28: [
+                {
+                    "to": 11,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            29: [
+                {
+                    "to": 14,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            30: [
+                {
+                    "to": 14,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            31: [
+                {
+                    "to": 16,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            32: [
+                {
+                    "to": 17,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            33: [
+                {
+                    "to": 18,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            34: [
+                {
+                    "to": 19,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            35: [
+                {
+                    "to": 19,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            36: [
+                {
+                    "to": 19,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            37: [
+                {
+                    "to": 21,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            38: [
+                {
+                    "to": 21,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            39: [
+                {
+                    "to": 22,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            40: [
+                {
+                    "to": 22,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+        }
+
+    @pytest.fixture
+    def heme_2d_atoms_data(self):
+        return {
+            1: {
+                "coords": np.array([19.8676, -19.1177, -1.5852]),
+                "element": "O",
+                "bond_to": {41: "C"},
+            },
+            2: {
+                "coords": np.array([21.6176, -17.7372, -1.5534]),
+                "element": "O",
+                "bond_to": {41: "C"},
+            },
+            3: {
+                "coords": np.array([27.7926, -11.3449, -1.3889]),
+                "element": "O",
+                "bond_to": {42: "C"},
+            },
+            4: {
+                "coords": np.array([26.2694, -12.9697, -1.4323]),
+                "element": "O",
+                "bond_to": {42: "C"},
+            },
+            5: {
+                "coords": np.array([18.712, -11.9403, -1.2028]),
+                "element": "N",
+                "bond_to": {9: "C", 11: "C"},
+            },
+            6: {
+                "coords": np.array([20.6948, -9.9176, -1.242]),
+                "element": "N",
+                "bond_to": {10: "C", 12: "C"},
+            },
+            7: {
+                "coords": np.array([16.6792, -9.9261, -1.2508]),
+                "element": "N",
+                "bond_to": {13: "C", 16: "C"},
+            },
+            8: {
+                "coords": np.array([18.652, -7.914, -1.2977]),
+                "element": "N",
+                "bond_to": {14: "C", 15: "C"},
+            },
+            9: {
+                "coords": np.array([19.7942, -12.742, -1.1871]),
+                "element": "C",
+                "bond_to": {5: "N", 17: "C", 18: "C"},
+            },
+            10: {
+                "coords": np.array([21.5042, -10.9948, -1.212]),
+                "element": "C",
+                "bond_to": {6: "N", 17: "C", 19: "C"},
+            },
+            11: {
+                "coords": np.array([17.6235, -12.7312, -1.1933]),
+                "element": "C",
+                "bond_to": {5: "N", 20: "C", 22: "C"},
+            },
+            12: {
+                "coords": np.array([21.4739, -8.8181, -1.2818]),
+                "element": "C",
+                "bond_to": {6: "N", 21: "C", 23: "C"},
+            },
+            13: {
+                "coords": np.array([15.8963, -11.0195, -1.2741]),
+                "element": "C",
+                "bond_to": {7: "N", 20: "C", 25: "C"},
+            },
+            14: {
+                "coords": np.array([19.7297, -7.1105, -1.3373]),
+                "element": "C",
+                "bond_to": {8: "N", 21: "C", 27: "C"},
+            },
+            15: {
+                "coords": np.array([17.5621, -7.1222, -1.3162]),
+                "element": "C",
+                "bond_to": {8: "N", 24: "C", 26: "C"},
+            },
+            16: {
+                "coords": np.array([15.8641, -8.8548, -1.3049]),
+                "element": "C",
+                "bond_to": {7: "N", 24: "C", 28: "C"},
+            },
+            17: {
+                "coords": np.array([21.1173, -12.3295, -1.1815]),
+                "element": "C",
+                "bond_to": {9: "C", 10: "C", 43: "H"},
+            },
+            18: {
+                "coords": np.array([19.3583, -14.1568, -1.1795]),
+                "element": "C",
+                "bond_to": {9: "C", 22: "C", 29: "C"},
+            },
+            19: {
+                "coords": np.array([22.8218, -10.5756, -1.2277]),
+                "element": "C",
+                "bond_to": {10: "C", 23: "C", 30: "C"},
+            },
+            20: {
+                "coords": np.array([16.3017, -12.3365, -1.2347]),
+                "element": "C",
+                "bond_to": {11: "C", 13: "C", 44: "H"},
+            },
+            21: {
+                "coords": np.array([21.0557, -7.4979, -1.3319]),
+                "element": "C",
+                "bond_to": {12: "C", 14: "C", 45: "H"},
+            },
+            22: {
+                "coords": np.array([18.0198, -14.1321, -1.1515]),
+                "element": "C",
+                "bond_to": {11: "C", 18: "C", 31: "C"},
+            },
+            23: {
+                "coords": np.array([22.8005, -9.1973, -1.2723]),
+                "element": "C",
+                "bond_to": {12: "C", 19: "C", 32: "C"},
+            },
+            24: {
+                "coords": np.array([16.2416, -7.528, -1.3195]),
+                "element": "C",
+                "bond_to": {15: "C", 16: "C", 46: "H"},
+            },
+            25: {
+                "coords": np.array([14.574, -10.6443, -1.3497]),
+                "element": "C",
+                "bond_to": {13: "C", 28: "C", 33: "C"},
+            },
+            26: {
+                "coords": np.array([17.9666, -5.7197, -1.362]),
+                "element": "C",
+                "bond_to": {15: "C", 27: "C", 34: "C"},
+            },
+            27: {
+                "coords": np.array([19.3008, -5.7093, -1.3935]),
+                "element": "C",
+                "bond_to": {14: "C", 26: "C", 35: "C"},
+            },
+            28: {
+                "coords": np.array([14.5546, -9.2737, -1.3631]),
+                "element": "C",
+                "bond_to": {16: "C", 25: "C", 36: "C"},
+            },
+            29: {
+                "coords": np.array([20.2863, -15.374, -1.1968]),
+                "element": "C",
+                "bond_to": {18: "C", 37: "C", 47: "H", 48: "H"},
+            },
+            30: {
+                "coords": np.array([24.0278, -11.5107, -1.2136]),
+                "element": "C",
+                "bond_to": {19: "C", 38: "C", 49: "H", 50: "H"},
+            },
+            31: {
+                "coords": np.array([17.0435, -15.2964, -1.0919]),
+                "element": "C",
+                "bond_to": {22: "C", 51: "H", 52: "H", 53: "H"},
+            },
+            32: {
+                "coords": np.array([23.9858, -8.2439, -1.3163]),
+                "element": "C",
+                "bond_to": {23: "C", 54: "H", 55: "H", 56: "H"},
+            },
+            33: {
+                "coords": np.array([13.3797, -11.5726, -1.4182]),
+                "element": "C",
+                "bond_to": {25: "C", 57: "H", 58: "H", 59: "H"},
+            },
+            34: {
+                "coords": np.array([17.0243, -4.5359, -1.4015]),
+                "element": "C",
+                "bond_to": {26: "C", 60: "H", 61: "H", 62: "H"},
+            },
+            35: {
+                "coords": np.array([20.2129, -4.5454, -1.5006]),
+                "element": "C",
+                "bond_to": {27: "C", 39: "C", 63: "H"},
+            },
+            36: {
+                "coords": np.array([13.3706, -8.401, -1.4492]),
+                "element": "C",
+                "bond_to": {28: "C", 40: "C", 64: "H"},
+            },
+            37: {
+                "coords": np.array([19.5448, -16.7096, -1.1026]),
+                "element": "C",
+                "bond_to": {29: "C", 41: "C", 65: "H", 66: "H"},
+            },
+            38: {
+                "coords": np.array([25.3819, -10.8068, -1.1561]),
+                "element": "C",
+                "bond_to": {30: "C", 42: "C", 67: "H", 68: "H"},
+            },
+            39: {
+                "coords": np.array([20.0311, -3.4198, -0.7937]),
+                "element": "C",
+                "bond_to": {35: "C", 69: "H", 70: "H"},
+            },
+            40: {
+                "coords": np.array([12.2878, -8.6418, -0.6967]),
+                "element": "C",
+                "bond_to": {36: "C", 71: "H", 72: "H"},
+            },
+            41: {
+                "coords": np.array([20.4233, -17.8843, -1.4349]),
+                "element": "C",
+                "bond_to": {1: "O", 2: "O", 37: "C"},
+            },
+            42: {
+                "coords": np.array([26.5085, -11.7886, -1.3281]),
+                "element": "C",
+                "bond_to": {3: "O", 4: "O", 38: "C"},
+            },
+            43: {
+                "coords": np.array([21.9066, -13.1092, -1.1619]),
+                "element": "H",
+                "bond_to": {17: "C"},
+            },
+            44: {
+                "coords": np.array([15.5193, -13.1211, -1.2439]),
+                "element": "H",
+                "bond_to": {20: "C"},
+            },
+            45: {
+                "coords": np.array([21.8243, -6.6991, -1.3699]),
+                "element": "H",
+                "bond_to": {21: "C"},
+            },
+            46: {
+                "coords": np.array([15.4484, -6.7533, -1.3538]),
+                "element": "H",
+                "bond_to": {24: "C"},
+            },
+            47: {
+                "coords": np.array([20.853, -15.3637, -2.1563]),
+                "element": "H",
+                "bond_to": {29: "C"},
+            },
+            48: {
+                "coords": np.array([20.9975, -15.2995, -0.3446]),
+                "element": "H",
+                "bond_to": {29: "C"},
+            },
+            49: {
+                "coords": np.array([23.9479, -12.1771, -0.3268]),
+                "element": "H",
+                "bond_to": {30: "C"},
+            },
+            50: {
+                "coords": np.array([23.9942, -12.1286, -2.1377]),
+                "element": "H",
+                "bond_to": {30: "C"},
+            },
+            51: {
+                "coords": np.array([16.0463, -14.9562, -0.7387]),
+                "element": "H",
+                "bond_to": {31: "C"},
+            },
+            52: {
+                "coords": np.array([16.9377, -15.726, -2.1108]),
+                "element": "H",
+                "bond_to": {31: "C"},
+            },
+            53: {
+                "coords": np.array([17.4003, -16.0738, -0.3855]),
+                "element": "H",
+                "bond_to": {31: "C"},
+            },
+            54: {
+                "coords": np.array([24.7924, -8.5743, -0.6292]),
+                "element": "H",
+                "bond_to": {32: "C"},
+            },
+            55: {
+                "coords": np.array([24.3704, -8.2067, -2.3585]),
+                "element": "H",
+                "bond_to": {32: "C"},
+            },
+            56: {
+                "coords": np.array([23.6777, -7.2237, -1.0041]),
+                "element": "H",
+                "bond_to": {32: "C"},
+            },
+            57: {
+                "coords": np.array([12.5446, -11.0758, -1.9545]),
+                "element": "H",
+                "bond_to": {33: "C"},
+            },
+            58: {
+                "coords": np.array([13.0626, -11.8208, -0.3832]),
+                "element": "H",
+                "bond_to": {33: "C"},
+            },
+            59: {
+                "coords": np.array([13.6319, -12.5121, -1.9593]),
+                "element": "H",
+                "bond_to": {33: "C"},
+            },
+            60: {
+                "coords": np.array([17.4405, -3.6831, -0.8297]),
+                "element": "H",
+                "bond_to": {34: "C"},
+            },
+            61: {
+                "coords": np.array([16.8732, -4.2338, -2.4592]),
+                "element": "H",
+                "bond_to": {34: "C"},
+            },
+            62: {
+                "coords": np.array([16.0435, -4.8051, -0.9507]),
+                "element": "H",
+                "bond_to": {34: "C"},
+            },
+            63: {
+                "coords": np.array([21.0809, -4.6133, -2.1892]),
+                "element": "H",
+                "bond_to": {35: "C"},
+            },
+            64: {
+                "coords": np.array([13.3792, -7.54, -2.1494]),
+                "element": "H",
+                "bond_to": {36: "C"},
+            },
+            65: {
+                "coords": np.array([18.7066, -16.6943, -1.832]),
+                "element": "H",
+                "bond_to": {37: "C"},
+            },
+            66: {
+                "coords": np.array([19.1517, -16.8412, -0.071]),
+                "element": "H",
+                "bond_to": {37: "C"},
+            },
+            67: {
+                "coords": np.array([25.4896, -10.3254, -0.1583]),
+                "element": "H",
+                "bond_to": {38: "C"},
+            },
+            68: {
+                "coords": np.array([25.4419, -10.0395, -1.9593]),
+                "element": "H",
+                "bond_to": {38: "C"},
+            },
+            69: {
+                "coords": np.array([19.1774, -3.3276, -0.0927]),
+                "element": "H",
+                "bond_to": {39: "C"},
+            },
+            70: {
+                "coords": np.array([20.7461, -2.5757, -0.9082]),
+                "element": "H",
+                "bond_to": {39: "C"},
+            },
+            71: {
+                "coords": np.array([12.282, -9.5056, 0.0]),
+                "element": "H",
+                "bond_to": {40: "C"},
+            },
+            72: {
+                "coords": np.array([11.3978, -7.9789, -0.7703]),
+                "element": "H",
+                "bond_to": {40: "C"},
+            },
+            73: {"coords": np.array([18.5078, -10.0631, 0.0]), "element": "Fe"},
+        }
+
+    @pytest.fixture
+    def heme_2d_bonds_data(self):
+        return {
+            1: [
+                {
+                    "to": 41,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            2: [
+                {
+                    "to": 41,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            3: [
+                {
+                    "to": 42,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            4: [
+                {
+                    "to": 42,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            5: [
+                {
+                    "to": 9,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 11,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            6: [
+                {
+                    "to": 10,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 12,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            7: [
+                {
+                    "to": 13,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 16,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            8: [
+                {
+                    "to": 14,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 15,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            9: [
+                {
+                    "to": 17,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 18,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            10: [
+                {
+                    "to": 17,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 19,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            11: [
+                {
+                    "to": 20,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 22,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            12: [
+                {
+                    "to": 21,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 23,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            13: [
+                {
+                    "to": 20,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 25,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            14: [
+                {
+                    "to": 21,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 27,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            15: [
+                {
+                    "to": 24,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 26,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            16: [
+                {
+                    "to": 24,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 28,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            17: [
+                {
+                    "to": 43,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            18: [
+                {
+                    "to": 22,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 29,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            19: [
+                {
+                    "to": 23,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 30,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            20: [
+                {
+                    "to": 44,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            21: [
+                {
+                    "to": 45,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            22: [
+                {
+                    "to": 31,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            23: [
+                {
+                    "to": 32,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            24: [
+                {
+                    "to": 46,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            25: [
+                {
+                    "to": 28,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 33,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            26: [
+                {
+                    "to": 27,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 34,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            27: [
+                {
+                    "to": 35,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            28: [
+                {
+                    "to": 36,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            29: [
+                {
+                    "to": 37,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 47,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 48,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            30: [
+                {
+                    "to": 38,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 49,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 50,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            31: [
+                {
+                    "to": 51,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 52,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 53,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            32: [
+                {
+                    "to": 54,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 55,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 56,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            33: [
+                {
+                    "to": 57,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 58,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 59,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            34: [
+                {
+                    "to": 60,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 61,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 62,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            35: [
+                {
+                    "to": 39,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 63,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            36: [
+                {
+                    "to": 40,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 64,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            37: [
+                {
+                    "to": 41,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 65,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 66,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            38: [
+                {
+                    "to": 42,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 67,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 68,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            39: [
+                {
+                    "to": 69,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 70,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            40: [
+                {
+                    "to": 71,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 72,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+        }
+
+    @pytest.fixture
+    def heme_group_atoms_data(self):
+        return {
+            1: {"coords": np.array([18.856, -10.1645, 0.0]), "element": "Fe"},
+            2: {
+                "coords": np.array([19.4111, -18.3615, 0.0]),
+                "element": "O",
+                "bond_to": {42: "C"},
+            },
+            3: {
+                "coords": np.array([21.4059, -17.2096, 0.0]),
+                "element": "O",
+                "bond_to": {42: "C"},
+            },
+            4: {
+                "coords": np.array([27.1847, -11.2, 0.0]),
+                "element": "O",
+                "bond_to": {43: "C"},
+            },
+            5: {
+                "coords": np.array([25.7979, -13.0395, 0.0]),
+                "element": "O",
+                "bond_to": {43: "C"},
+            },
+            6: {
+                "coords": np.array([18.7949, -11.5492, 0.0]),
+                "element": "N",
+                "bond_to": {10: "C", 12: "C"},
+            },
+            7: {
+                "coords": np.array([20.5798, -10.0015, 0.0]),
+                "element": "N",
+                "bond_to": {11: "C", 13: "C"},
+            },
+            8: {
+                "coords": np.array([16.9864, -10.0506, 0.0]),
+                "element": "N",
+                "bond_to": {14: "C", 17: "C"},
+            },
+            9: {
+                "coords": np.array([18.7584, -8.4699, 0.0]),
+                "element": "N",
+                "bond_to": {15: "C", 16: "C"},
+            },
+            10: {
+                "coords": np.array([19.7909, -12.5337, 0.0]),
+                "element": "C",
+                "bond_to": {6: "N", 18: "C", 19: "C"},
+            },
+            11: {
+                "coords": np.array([21.4317, -10.9754, 0.0]),
+                "element": "C",
+                "bond_to": {7: "N", 18: "C", 20: "C"},
+            },
+            12: {
+                "coords": np.array([17.7681, -12.576, 0.0]),
+                "element": "C",
+                "bond_to": {6: "N", 21: "C", 23: "C"},
+            },
+            13: {
+                "coords": np.array([21.3894, -9.037, 0.0]),
+                "element": "C",
+                "bond_to": {7: "N", 22: "C", 24: "C"},
+            },
+            14: {
+                "coords": np.array([16.1676, -11.0177, 0.0]),
+                "element": "C",
+                "bond_to": {8: "N", 21: "C", 26: "C"},
+            },
+            15: {
+                "coords": np.array([19.7909, -7.4787, 0.0]),
+                "element": "C",
+                "bond_to": {9: "N", 22: "C", 28: "C"},
+            },
+            16: {
+                "coords": np.array([17.7258, -7.521, 0.0]),
+                "element": "C",
+                "bond_to": {9: "N", 25: "C", 27: "C"},
+            },
+            17: {
+                "coords": np.array([16.1676, -9.0793, 0.0]),
+                "element": "C",
+                "bond_to": {8: "N", 25: "C", 29: "C"},
+            },
+            18: {
+                "coords": np.array([21.0536, -12.1959, 0.0]),
+                "element": "C",
+                "bond_to": {10: "C", 11: "C", 44: "H"},
+            },
+            19: {
+                "coords": np.array([19.4109, -13.7543, 0.0]),
+                "element": "C",
+                "bond_to": {10: "C", 23: "C", 30: "C"},
+            },
+            20: {
+                "coords": np.array([22.612, -10.6375, 0.0]),
+                "element": "C",
+                "bond_to": {11: "C", 24: "C", 31: "C"},
+            },
+            21: {
+                "coords": np.array([16.5476, -12.1959, 0.0]),
+                "element": "C",
+                "bond_to": {12: "C", 14: "C", 45: "H"},
+            },
+            22: {
+                "coords": np.array([21.0115, -7.8588, 0.0]),
+                "element": "C",
+                "bond_to": {13: "C", 15: "C", 46: "H"},
+            },
+            23: {
+                "coords": np.array([18.1482, -13.7543, 0.0]),
+                "element": "C",
+                "bond_to": {12: "C", 19: "C", 32: "C"},
+            },
+            24: {
+                "coords": np.array([22.612, -9.3748, 0.0]),
+                "element": "C",
+                "bond_to": {13: "C", 20: "C", 33: "C"},
+            },
+            25: {
+                "coords": np.array([16.5054, -7.8588, 0.0]),
+                "element": "C",
+                "bond_to": {16: "C", 17: "C", 47: "H"},
+            },
+            26: {
+                "coords": np.array([14.9472, -10.6797, 0.0]),
+                "element": "C",
+                "bond_to": {14: "C", 29: "C", 34: "C"},
+            },
+            27: {
+                "coords": np.array([18.1059, -6.3004, 0.0]),
+                "element": "C",
+                "bond_to": {16: "C", 28: "C", 35: "C"},
+            },
+            28: {
+                "coords": np.array([19.3686, -6.3004, 0.0]),
+                "element": "C",
+                "bond_to": {15: "C", 27: "C", 36: "C"},
+            },
+            29: {
+                "coords": np.array([14.9472, -9.4171, 0.0]),
+                "element": "C",
+                "bond_to": {17: "C", 26: "C", 37: "C"},
+            },
+            30: {
+                "coords": np.array([20.0759, -14.906, 0.0]),
+                "element": "C",
+                "bond_to": {19: "C", 38: "C", 48: "H", 49: "H"},
+            },
+            31: {
+                "coords": np.array([23.6739, -11.4382, 0.0]),
+                "element": "C",
+                "bond_to": {20: "C", 39: "C", 50: "H", 51: "H"},
+            },
+            32: {
+                "coords": np.array([17.3654, -14.8294, 0.0]),
+                "element": "C",
+                "bond_to": {23: "C", 52: "H", 53: "H", 54: "H"},
+            },
+            33: {
+                "coords": np.array([23.6703, -8.5694, 0.0]),
+                "element": "C",
+                "bond_to": {24: "C", 55: "H", 56: "H", 57: "H"},
+            },
+            34: {
+                "coords": np.array([13.8886, -11.485, 0.0]),
+                "element": "C",
+                "bond_to": {26: "C", 58: "H", 59: "H", 60: "H"},
+            },
+            35: {
+                "coords": np.array([17.3176, -5.2292, 0.0]),
+                "element": "C",
+                "bond_to": {27: "C", 61: "H", 62: "H", 63: "H"},
+            },
+            36: {
+                "coords": np.array([20.1343, -5.2129, 0.0]),
+                "element": "C",
+                "bond_to": {28: "C", 40: "C", 64: "H"},
+            },
+            37: {
+                "coords": np.array([13.8886, -8.6118, 0.0]),
+                "element": "C",
+                "bond_to": {29: "C", 41: "C", 65: "H"},
+            },
+            38: {
+                "coords": np.array([19.4109, -16.0578, 0.0]),
+                "element": "C",
+                "bond_to": {30: "C", 42: "C", 66: "H", 67: "H"},
+            },
+            39: {
+                "coords": np.array([24.8983, -10.9187, 0.0]),
+                "element": "C",
+                "bond_to": {31: "C", 43: "C", 68: "H", 69: "H"},
+            },
+            40: {
+                "coords": np.array([19.5752, -4.006, 0.0]),
+                "element": "C",
+                "bond_to": {36: "C", 70: "H", 71: "H"},
+            },
+            41: {
+                "coords": np.array([12.662, -9.1259, 0.0]),
+                "element": "C",
+                "bond_to": {37: "C", 72: "H", 73: "H"},
+            },
+            42: {
+                "coords": np.array([20.0759, -17.2096, 0.0]),
+                "element": "C",
+                "bond_to": {2: "O", 3: "O", 38: "C"},
+            },
+            43: {
+                "coords": np.array([25.9603, -11.7195, 0.0]),
+                "element": "C",
+                "bond_to": {4: "O", 5: "O", 39: "C"},
+            },
+            44: {
+                "coords": np.array([21.6253, -12.7901, 0.0]),
+                "element": "H",
+                "bond_to": {18: "C"},
+            },
+            45: {
+                "coords": np.array([15.9674, -12.7819, 0.0]),
+                "element": "H",
+                "bond_to": {21: "C"},
+            },
+            46: {
+                "coords": np.array([21.592, -7.2733, 0.0]),
+                "element": "H",
+                "bond_to": {22: "C"},
+            },
+            47: {
+                "coords": np.array([15.9223, -7.2757, 0.0]),
+                "element": "H",
+                "bond_to": {25: "C"},
+            },
+            48: {
+                "coords": np.array([20.7077, -15.4361, 0.0]),
+                "element": "H",
+                "bond_to": {30: "C"},
+            },
+            49: {
+                "coords": np.array([20.7077, -14.376, 0.0]),
+                "element": "H",
+                "bond_to": {30: "C"},
+            },
+            50: {
+                "coords": np.array([23.0707, -12.0004, 0.0]),
+                "element": "H",
+                "bond_to": {31: "C"},
+            },
+            51: {
+                "coords": np.array([24.1229, -12.1298, 0.0]),
+                "element": "H",
+                "bond_to": {31: "C"},
+            },
+            52: {
+                "coords": np.array([16.6988, -14.344, 0.0]),
+                "element": "H",
+                "bond_to": {32: "C"},
+            },
+            53: {
+                "coords": np.array([16.8799, -15.496, 0.0]),
+                "element": "H",
+                "bond_to": {32: "C"},
+            },
+            54: {
+                "coords": np.array([18.032, -15.3148, 0.0]),
+                "element": "H",
+                "bond_to": {32: "C"},
+            },
+            55: {
+                "coords": np.array([24.1697, -9.2256, 0.0]),
+                "element": "H",
+                "bond_to": {33: "C"},
+            },
+            56: {
+                "coords": np.array([24.3265, -8.07, 0.0]),
+                "element": "H",
+                "bond_to": {33: "C"},
+            },
+            57: {
+                "coords": np.array([23.1708, -7.9133, 0.0]),
+                "element": "H",
+                "bond_to": {33: "C"},
+            },
+            58: {
+                "coords": np.array([13.3894, -10.8288, 0.0]),
+                "element": "H",
+                "bond_to": {34: "C"},
+            },
+            59: {
+                "coords": np.array([13.2323, -11.9843, 0.0]),
+                "element": "H",
+                "bond_to": {34: "C"},
+            },
+            60: {
+                "coords": np.array([14.3879, -12.1412, 0.0]),
+                "element": "H",
+                "bond_to": {34: "C"},
+            },
+            61: {
+                "coords": np.array([17.9819, -4.7405, 0.0]),
+                "element": "H",
+                "bond_to": {35: "C"},
+            },
+            62: {
+                "coords": np.array([16.8289, -4.565, 0.0]),
+                "element": "H",
+                "bond_to": {35: "C"},
+            },
+            63: {
+                "coords": np.array([16.6536, -5.718, 0.0]),
+                "element": "H",
+                "bond_to": {35: "C"},
+            },
+            64: {
+                "coords": np.array([20.9556, -5.2868, 0.0]),
+                "element": "H",
+                "bond_to": {36: "C"},
+            },
+            65: {
+                "coords": np.array([13.9929, -7.7939, 0.0]),
+                "element": "H",
+                "bond_to": {37: "C"},
+            },
+            66: {
+                "coords": np.array([18.7793, -15.5278, 0.0]),
+                "element": "H",
+                "bond_to": {38: "C"},
+            },
+            67: {
+                "coords": np.array([18.7793, -16.5878, 0.0]),
+                "element": "H",
+                "bond_to": {38: "C"},
+            },
+            68: {
+                "coords": np.array([25.5015, -10.3565, 0.0]),
+                "element": "H",
+                "bond_to": {39: "C"},
+            },
+            69: {
+                "coords": np.array([24.4494, -10.2271, 0.0]),
+                "element": "H",
+                "bond_to": {39: "C"},
+            },
+            70: {
+                "coords": np.array([20.0499, -3.3318, 0.0]),
+                "element": "H",
+                "bond_to": {40: "C"},
+            },
+            71: {
+                "coords": np.array([18.7539, -3.9322, 0.0]),
+                "element": "H",
+                "bond_to": {40: "C"},
+            },
+            72: {
+                "coords": np.array([12.0058, -8.6267, 0.0]),
+                "element": "H",
+                "bond_to": {41: "C"},
+            },
+            73: {
+                "coords": np.array([12.5577, -9.9439, 0.0]),
+                "element": "H",
+                "bond_to": {41: "C"},
+            },
+        }
+
+    @pytest.fixture
+    def heme_group_bonds_data(self):
+        return {
+            2: [
+                {
+                    "to": 42,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            3: [
+                {
+                    "to": 42,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            4: [
+                {
+                    "to": 43,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            5: [
+                {
+                    "to": 43,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            6: [
+                {
+                    "to": 10,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 12,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            7: [
+                {
+                    "to": 11,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 13,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            8: [
+                {
+                    "to": 14,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 17,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            9: [
+                {
+                    "to": 15,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 16,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            10: [
+                {
+                    "to": 18,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 19,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            11: [
+                {
+                    "to": 18,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 20,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            12: [
+                {
+                    "to": 21,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 23,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            13: [
+                {
+                    "to": 22,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 24,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            14: [
+                {
+                    "to": 21,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 26,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            15: [
+                {
+                    "to": 22,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 28,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            16: [
+                {
+                    "to": 25,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 27,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            17: [
+                {
+                    "to": 25,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 29,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            18: [
+                {
+                    "to": 44,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            19: [
+                {
+                    "to": 23,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 30,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            20: [
+                {
+                    "to": 24,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 31,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            21: [
+                {
+                    "to": 45,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            22: [
+                {
+                    "to": 46,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            23: [
+                {
+                    "to": 32,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            24: [
+                {
+                    "to": 33,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            25: [
+                {
+                    "to": 47,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            26: [
+                {
+                    "to": 29,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 34,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            27: [
+                {
+                    "to": 28,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 35,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            28: [
+                {
+                    "to": 36,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            29: [
+                {
+                    "to": 37,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                }
+            ],
+            30: [
+                {
+                    "to": 38,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 48,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 49,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            31: [
+                {
+                    "to": 39,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 50,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 51,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            32: [
+                {
+                    "to": 52,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 53,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 54,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            33: [
+                {
+                    "to": 55,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 56,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 57,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            34: [
+                {
+                    "to": 58,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 59,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 60,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            35: [
+                {
+                    "to": 61,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 62,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 63,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            36: [
+                {
+                    "to": 40,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 64,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            37: [
+                {
+                    "to": 41,
+                    "type": "2",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 65,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            38: [
+                {
+                    "to": 42,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 66,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 67,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            39: [
+                {
+                    "to": 43,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 68,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 69,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            40: [
+                {
+                    "to": 70,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 71,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+            41: [
+                {
+                    "to": 72,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+                {
+                    "to": 73,
+                    "type": "1",
+                    "stereo": 0,
+                    "topology": 0,
+                    "reacting_center_status": 0,
+                },
+            ],
+        }
+
+    def compare_read_molecule_and_known_values(
+        self, read_data, known_atoms_data, known_bonds_data
+    ):
+        # Test atoms data
+        atoms_data = read_data.atoms_data
+        assert isinstance(atoms_data, dict)
+        assert known_atoms_data.keys() == atoms_data.keys()
+        for atom_index, atom_data in atoms_data.items():
+            assert atom_index in known_atoms_data
+            assert np.array_equal(
+                atom_data["coords"], known_atoms_data[atom_index]["coords"]
+            )
+            assert atom_data["element"] == known_atoms_data[atom_index]["element"]
+            if atom_data.get("bond_to"):
+                assert atom_data["bond_to"] == known_atoms_data[atom_index]["bond_to"]
+
+        assert read_data.bonds_data == known_bonds_data
+
+    def test_read_files(
+        self,
+        morphine_file_data,
+        morphine_3d_file_data,
+        heme_2d_file_data,
+        heme_group_file_data,
+    ):
+        # Test read 2D morphine file
+        assert morphine_file_data == MolParser.read_file(self.morphine_mol_file)
+
+        # Test read morphine 3D file
+        assert morphine_3d_file_data == MolParser.read_file(self.morphine_3d_mol_file)
+
+        # Test read heme 2d
+        assert heme_2d_file_data == MolParser.read_file(self.heme_2d_mol_file)
+
+        # Test read heme 3d
+        assert heme_group_file_data == MolParser.read_file(self.heme_group_mol_file)
+
+    def test_parse_files(
+        self,
+        morphine_atoms_data,
+        morphine_bonds_data,
+        morphine_3d_atoms_data,
+        morphine_3d_bonds_data,
+        heme_2d_atoms_data,
+        heme_2d_bonds_data,
+        heme_group_atoms_data,
+        heme_group_bonds_data,
+    ):
+        # Test parse 2D morphine file
+        self.compare_read_molecule_and_known_values(
+            MolParser(self.morphine_mol_file), morphine_atoms_data, morphine_bonds_data
+        )
+
+        # Test parse 3D morphine file
+        self.compare_read_molecule_and_known_values(
+            MolParser(self.morphine_3d_mol_file),
+            morphine_3d_atoms_data,
+            morphine_3d_bonds_data,
+        )
+
+        # Test parse 2D Heme group
+        self.compare_read_molecule_and_known_values(
+            MolParser(self.heme_2d_mol_file), heme_2d_atoms_data, heme_2d_bonds_data
+        )
+
+        # Test parse 3D heme group
+        self.compare_read_molecule_and_known_values(
+            MolParser(self.heme_group_mol_file),
+            heme_group_atoms_data,
+            heme_group_bonds_data,
+        )
